backend/requirements.txt
flask
flask-cors
python-dotenv
ibm-watson
